[hm] Switching back to int alarm compares.
Low: Arming point = Thresh + 1.0f, Trigger point = Thresh - 0.1f
High: Arming point = Thresh - 1.1f, Trigger point = Thresh

+++ arduino/heatermeter/grillpid.h
@@ -46,7 +46,7 @@ public:
   ProbeAlarm(void) {};
 
   // Check value against Low/High
-  void updateStatus(float value);
+  void updateStatus(int value);
   void setLow(int value);
   void setHigh(int value);
   int getLow(void) const { return Thresholds[ALARM_IDX_LOW]; }
